requirements.txt
python-telegram-bot==20.3
Flask==3.0.0
pymongo==4.5.0
gunicorn==21.2.0